core folder created, then App, GlobalStyle, rootSaga, store and theme moved to core
