+++ src/core/App.js
@@ -1,11 +1,11 @@
 import { HashRouter, Redirect, Route, Switch } from "react-router-dom";
-import { Header } from "./common/Header";
-import MoviesListPage from "./features/movies/MoviesListPage";
-import PeopleListPage from "./features/people/PeopleListPage";
-import ProfilePage from "./features/people/ProfilePage";
-import MoviePage from "./features/movies/MoviePage";
+import { Header } from "../common/Header";
+import MoviesListPage from "../features/movies/MoviesListPage";
+import PeopleListPage from "../features/people/PeopleListPage";
+import ProfilePage from "../features/people/ProfilePage";
+import MoviePage from "../features/movies/MoviePage";
 import { useDispatch } from "react-redux";
-import { fetchGenres } from "./features/genresSlice";
+import { fetchGenres } from "../features/genresSlice";
 import { useEffect } from "react";
 
 const App = () => {
