+++ src/core/store.js
@@ -1,10 +1,10 @@
 import { configureStore } from "@reduxjs/toolkit";
 import createSagaMiddleware from "redux-saga";
-import { genresReducer } from "./features/genresSlice";
-import { movieReducer } from "./features/movies/MoviePage/movieSlice";
-import { moviesReducer } from "./features/movies/MoviesListPage/moviesSlice";
-import { peopleReducer } from "./features/people/peopleSlice";
-import { personReducer } from "./features/people/ProfilePage/personSlice";
+import { genresReducer } from "../features/genresSlice";
+import { movieReducer } from "../features/movies/MoviePage/movieSlice";
+import { moviesReducer } from "../features/movies/MoviesListPage/moviesSlice";
+import { peopleReducer } from "../features/people/peopleSlice";
+import { personReducer } from "../features/people/ProfilePage/personSlice";
 import rootSaga from "./rootSaga";
 
 const sagaMiddleware = createSagaMiddleware();
