+++ src/index.js
@@ -1,12 +1,12 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import { ThemeProvider } from "styled-components";
-import App from "./App";
+import App from "../src/core/App";
 import reportWebVitals from "./reportWebVitals";
-import { theme } from "./theme";
-import { GlobalStyle } from "./GlobalStyle";
+import { theme } from "../src/core/theme";
+import { GlobalStyle } from "../src/core/GlobalStyle";
 import { Provider } from "react-redux";
-import store from "./store";
+import store from "../src/core/store";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
